+++ history.md
@@ -1,3 +1,10 @@
+# v0.2.0 / 2016-04-15
+
+* Introduced native `Promise` support
+* Moved to gulp for build tasks running
+* Moved to eslint for code lint
+* The first found match is applied when multiple environment override files are present
+
 # v0.1.6 / 2015-11-16
 
 * Fixed issue where `null` values in config caused type coercion to fail